Implement all needed services

+++ src/main/java/com/example/model/Person.java
@@ -8,17 +8,19 @@ import javax.persistence.*;
 @Entity
 public class Person {
 
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String surname;
 
-    public int getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
